# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Grondona API
  description: A Kotlin HTTP server for the PRODE Norteamérica 2026, built with Spring Boot.
  version: 1.0.0
  contact:
    name: Grondona API Support
    email: cristian.rana8@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/tournaments/{tournamentId}/groups:
    get:
      tags:
        - Groups
      summary: Get all groups for the given tournament.
      description: |
        Retrieve all groups. Optionally filter by name substring.
      operationId: getAllGroups
      security:
        - bearerAuth: [ ]
      parameters:
        - name: tournamentId
          in: path
          required: true
          description: UUID of the tournament
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          required: false
          description: Text substring to filter groups by name (case-insensitive)
          schema:
            type: string
          example: gen
        - name: joined
          in: query
          required: false
          description: Flag indicating if it should show only non-joined groups or not
          schema:
            type: boolean
          example: false
      responses:
        "200":
          description: Groups retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupResponse"
    post:
      tags:
        - Groups
      summary: Create a new group for the given tournament
      description: |
        Create a new group with the given name, privacy setting, and member limit.
      operationId: createGroup
      security:
        - bearerAuth: [ ]
      parameters:
        - name: tournamentId
          in: path
          required: true
          description: UUID of the tournament
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
            example:
              name: My Group
              is_private: false
              maxMembers: 20
      responses:
        "201":
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResponse"
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Group name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: 409
                error: Conflict
                message: Group name 'General' already exists
                timestamp: "2024-01-01T12:00:00"
                data:
                  field: name
                  rejectedValue: General
  /api/tournaments/{tournamentId}/groups/{groupId}/join:
    post:
      tags:
        - Groups
      summary: Join a tournament group
      description: |
        Authenticated user joins the specified group.
        Fails if the user is already a member or if the group is full.
      operationId: joinGroup
      security:
        - bearerAuth: [ ]
      parameters:
        - name: tournamentId
          in: path
          required: true
          description: UUID of the tournament
          schema:
            type: string
            format: uuid
        - name: groupId
          in: path
          required: true
          description: UUID of the group to join
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Joined group successfully
        '400':
          description: Already a member or group is full
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                alreadyMember:
                  summary: User is already a member
                  value:
                    status: 400
                    error: Bad Request
                    message: You are already member of this group
                groupFull:
                  summary: Group has reached max members
                  value:
                    status: 400
                    error: Bad Request
                    message: Group is full
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tournaments/{tournamentId}/groups/{groupId}/leave:
    delete:
      tags:
        - Groups
      summary: Leave a tournament group
      description: |
        Authenticated user leaves the specified group.
        Fails if the user is not a member of the group.
      operationId: leaveGroup
      security:
        - bearerAuth: [ ]
      parameters:
        - name: tournamentId
          in: path
          required: true
          description: UUID of the tournament
          schema:
            type: string
            format: uuid
        - name: groupId
          in: path
          required: true
          description: UUID of the group to leave
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Left group successfully
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found or user is not a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tournaments/{tournamentId}/groups/{groupId}:
    get:
      tags:
        - Groups
      summary: Get a group by ID
      description: |
        Retrieve a specific group by its UUID.
      operationId: getGroup
      security:
        - bearerAuth: [ ]
      parameters:
        - name: groupId
          in: path
          required: true
          description: UUID of the group
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Group retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResponse"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Groups
      summary: Update a group
      description: |
        Partially update a group. Only provided fields will be updated.
      operationId: updateGroup
      security:
        - bearerAuth: [ ]
      parameters:
        - name: groupId
          in: path
          required: true
          description: UUID of the group to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupRequest"
            example:
              name: Updated Name
              maxMembers: 30
      responses:
        "200":
          description: Group updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResponse"
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Group name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: 409
                error: Conflict
                message: Group name 'General' already exists
                timestamp: "2024-01-01T12:00:00"
                data:
                  field: name
                  rejectedValue: General
    delete:
      tags:
        - Groups
      summary: Delete a group
      description: |
        Delete a group by its UUID.
      operationId: deleteGroup
      security:
        - bearerAuth: [ ]
      parameters:
        - name: groupId
          in: path
          required: true
          description: UUID of the group to delete
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Group deleted successfully
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
    ErrorResponse:
      type: object
      required:
        - status
        - error
        - message
        - timestamp
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: Bad Request
        message:
          type: string
          description: Error message
          example: Invalid username or password
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-01T12:00:00"
        data:
          oneOf:
            - $ref: '#/components/schemas/ErrorConflictExtraData'
    ErrorConflictExtraData:
      type: object
      required:
        - field
        - rejected_value
      properties:
        field:
          type: string
          description: The field that caused the conflict
          example: username
        rejected_value:
          type: string
          description: The value that was rejected due to conflict
          example: johndoe
...
    CreateGroupRequest:
      type: object
      required:
        - name
        - max_members
      properties:
        name:
          type: string
          description: Unique group name
          maxLength: 100
          example: My Group
        is_private:
          type: boolean
          description: Whether the group is private (defaults to false)
          default: false
          example: false
        max_members:
          type: integer
          format: int16
          description: Maximum number of members allowed
          minimum: 1
          example: 20
    UpdateGroupRequest:
      type: object
      description: All fields are optional. Only provided fields will be updated.
      properties:
        name:
          type: string
          description: New group name (must be unique)
          maxLength: 100
          example: Updated Group Name
        is_private:
          type: boolean
          description: New privacy setting
          example: true
        max_members:
          type: integer
          format: int16
          description: New maximum members limit
          minimum: 1
          example: 30
    GroupResponse:
      type: object
      required:
        - id
        - tournament_id
        - name
        - private
        - max_members
      properties:
        id:
          type: string
          format: uuid
          description: Unique group identifier
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        tournament_id:
          type: string
          format: uuid
          description: Unique tournament identifier
          example: fa4e6142-20eb-43ac-a772-7cb58f409720
        name:
          type: string
          description: Group name
          example: General
        is_private:
          type: boolean
          description: Whether the group is private
          example: false
        max_members:
          type: integer
          format: int16
          description: Maximum number of members allowed
          example: 50
...